requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.60.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3